PID速度制御器追加・エンコーダ速度取得１次近似追加

+++ Winch_M1/Core/Inc/actuator/motor/ir2302_bridge.h
@@ -8,8 +8,6 @@
 #ifndef INC_ACTUATOR_MOTOR_IR2302_BRIDGE_H_
 #define INC_ACTUATOR_MOTOR_IR2302_BRIDGE_H_
 
-#include <stdint.h>
-
 void IR2302bridge_Init(void);
 
 void IR2302bridge_Enable(void);
@@ -18,16 +16,16 @@ void IR2302bridge_Disable(void);
 /**
  * @brief ハーフブリッジ1 Duty設定
  *
- * duty : 0～1000
+ * duty : 0.0 ～ 1.0
  */
-void IR2302Bridge_SetHB1Duty(uint16_t duty);
+void IR2302Bridge_SetHB1Duty(float duty);
 
 /**
  * @brief ハーフブリッジ2 Duty設定
  *
- * duty : 0～1000
+ * duty : 0.0 ～ 1.0
  */
-void IR2302Bridge_SetHB2Duty(uint16_t duty);
+void IR2302Bridge_SetHB2Duty(float duty);
 
 
 #endif /* INC_ACTUATOR_MOTOR_IR2302_BRIDGE_H_ */
